% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_cut.R
\name{laplace_cut}
\alias{laplace_cut}
\title{Laplace approximation of the cut posterior}
\usage{
laplace_cut(logposterior1, logposterior2, init_theta1, init_theta2)
}
\arguments{
\item{logposterior1}{: a function of theta1}

\item{logposterior2}{: a function of theta1 and theta2}

\item{init_theta1}{: initial point for theta1}

\item{init_theta2}{: initial point for theta2}
}
\value{
a list containing:
\itemize{
\item thetahat : the MAP estimate
\item asympvar : the asymptotic variance matrix of the Laplace approximation
}
}
\description{
Centered at the MAP and more or less using the Hessian of the log-density of the cut posterior
}
