% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwmrth.R
\name{rwmrth}
\alias{rwmrth}
\title{Random walk Metropolis-Rosenbluth-Teller-Hastings algorithm}
\usage{
rwmrth(target, tuning_parameters)
}
\arguments{
\item{target}{target distribution; it must be a list containing:
- dimension: dimension of the target
- density: function for evaluating the log target density; it must take as input
a matrix with nchains rows and target$dimension columns and return a vector of length nchains}

\item{tuning_parameters}{must be a list containing:
- niterations: number of iterations
- nchains: number of chains
- cov_proposal: initial proposal covariance matrix
- rinit: function for initializing the chains; it must take as input
the number of chains and return a matrix with nchains rows and target$dimension columns
- adaptation: adaptation frequency (if 0 no adaptation)}
}
\value{
a list containing:
- chains: a list of length nchains; each element is a matrix with niterations rows and target$dimension columns
- current_dtarget: vector of length nchains with the log target density values at the final states of the chains
- dtarget_history: matrix with niterations rows and nchains columns with the log target density values at each iteration
- naccepts: total number of accepted proposals across all chains
- cov_proposal: final proposal covariance matrix
}
\description{
with multiple chains, adaptive proposal
}
